@@ -5,7 +5,7 @@ description: A sample command-line application.
 # author: 95310 <953108373@qq.com>
 
 environment:
-  sdk: '>=2.4.0 <3.0.0'
+  sdk: '>=2.12.2 <3.0.0'
 
 #dependencies:
 #  path: ^1.6.0
