@@ -1,10 +1,10 @@
-void main() async{
+void main() async {
 //  Future.wait([fun1(), fun2()]).then((res) {
 //    print(res);
 //  });
 
-print((fun3()).runtimeType);
-print((fun4()).runtimeType);
+  print((fun3()).runtimeType);
+  print((fun4()).runtimeType);
 }
 
 Future<dynamic> fun1() async {
@@ -15,21 +15,21 @@ Future<dynamic> fun1() async {
 }
 
 Future<String> fun2() async {
-  Future.delayed(Duration(milliseconds: 2), (){
+  Future.delayed(Duration(milliseconds: 2), () {
     print('fun2');
   });
   return '222222';
 }
 
-Future<dynamic> fun3() async{
-  Future.delayed(Duration(milliseconds: 2), (){
+Future<dynamic> fun3() async {
+  Future.delayed(Duration(milliseconds: 2), () {
     print('fun3');
   });
   return await fun2();
 }
 
-Future<String> fun4() async{
-  Future.delayed(Duration(milliseconds: 2), (){
+Future<String> fun4() async {
+  Future.delayed(Duration(milliseconds: 2), () {
     print('fun4');
   });
   return fun2();
