@@ -1,4 +1,4 @@
-import 'package:dart_demo_learn/dart_demo_learn.dart' as dart_demo_learn;
+import 'package:dart_demo_learn/main.dart' as dart_demo_learn;
 
 void main() {
 //  print('Hello world: ${dart_demo_learn.calculate()}!');
